Changed the file design structure of item service
