+++ item_service/item/src/item_service_impl/controller/handler.rs
@@ -16,9 +16,8 @@ pub struct AppState {
 }
 
 pub fn create_item(req: &HttpRequest<AppState>) -> Result<Item,CustomError> {
+    unimplemented!();
 
-    let header = req.headers().get_all("");
-    let payload = Json::<ItemData>::extract(req);
 }
 
 pub fn get_item(data: State<AppState>,item_id:Path<i32>) -> Result<Item,CustomError> {
