+++ item_service/item/src/item_service_impl/models/item_summary.rs
