+++ item_service/item/src/item_service_impl/models/mod.rs
@@ -1,7 +1,5 @@
 pub mod item_data;
 pub mod item;
-pub mod location;
-pub mod item_summary;
 pub mod item_status;
 pub mod p_item;
 pub mod item_state;
