--- item_service/item_service_impl/.gitignore
@@ -1,4 +1,0 @@
-.idea
-target
-*.iml
-*.lock
